restructured liblx to match libsbml src directory
