+++ src/liblx/common/liblx-config.h
@@ -65,11 +65,11 @@
 #ifndef LIBSBXML_CONFIG_H
 #define LIBSBXML_CONFIG_H 1
 
-#include <liblx/xml/common/liblx-namespace.h>
+#include <liblx/common/liblx-namespace.h>
 
-#include <liblx/xml/common/libsbxml-config-common.h>
+#include <liblx/common/libsbxml-config-common.h>
 
-#include <liblx/xml/common/liblx-version.h>
+#include <liblx/common/liblx-version.h>
 
 
 #endif  /* LIBSBXML_CONFIG_H */
