+++ src/liblx/compress/CompressCommon.h
@@ -42,7 +42,7 @@
 #define CompressCommon_h
 
 #include <exception>
-#include <liblx/xml/common/extern.h>
+#include <liblx/common/extern.h>
 
 
 LIBLX_CPP_NAMESPACE_BEGIN
