+++ src/liblx/compress/InputDecompressor.cpp
@@ -46,15 +46,15 @@
 #include <algorithm>
 #include <cstring>
 
-#include <liblx/xml/compress/InputDecompressor.h>
+#include <liblx/compress/InputDecompressor.h>
 
 #ifdef USE_ZLIB
-#include <liblx/xml/compress/zfstream.h>
-#include <liblx/xml/compress/zipfstream.h>
+#include <liblx/compress/zfstream.h>
+#include <liblx/compress/zipfstream.h>
 #endif //USE_ZLIB
 
 #ifdef USE_BZ2
-#include <liblx/xml/compress/bzfstream.h>
+#include <liblx/compress/bzfstream.h>
 #endif //USE_BZ2
 
 using namespace std;
