+++ src/liblx/xml/LibXMLTranscode.h
@@ -43,7 +43,7 @@
 
 #include <string>
 #include <libxml/parser.h>
-#include <liblx/xml/common/extern.h>
+#include <liblx/common/extern.h>
 
 LIBLX_CPP_NAMESPACE_BEGIN
