+++ src/liblx/xml/XMLAttributes.h
@@ -227,8 +227,8 @@ foo:myattribute="7"
 #ifndef XMLAttributes_h
 #define XMLAttributes_h
 
-#include <liblx/xml/common/extern.h>
-#include <liblx/xml/common/liblxfwd.h>
+#include <liblx/common/extern.h>
+#include <liblx/common/liblxfwd.h>
 
 
 #ifdef __cplusplus
