+++ src/liblx/xml/XMLError.h
@@ -339,8 +339,8 @@
 #include <stdio.h>
 
 
-#include <liblx/xml/common/extern.h>
-#include <liblx/xml/common/liblxfwd.h>
+#include <liblx/common/extern.h>
+#include <liblx/common/liblxfwd.h>
 
 LIBLX_CPP_NAMESPACE_BEGIN
