+++ src/liblx/xml/XMLErrorLog.h
@@ -82,9 +82,9 @@
 #ifndef XMLErrorLog_h
 #define XMLErrorLog_h
 
-#include <liblx/xml/common/extern.h>
+#include <liblx/common/extern.h>
 #include <liblx/xml/XMLError.h>
-#include <liblx/xml/common/liblxfwd.h>
+#include <liblx/common/liblxfwd.h>
 
 
 #ifdef __cplusplus
