+++ src/liblx/xml/XMLLogOverride.h
@@ -46,7 +46,7 @@
 
 #ifdef __cplusplus
 
-#include <liblx/xml/common/extern.h>
+#include <liblx/common/extern.h>
 #include <liblx/xml/XMLError.h>
 
 LIBLX_CPP_NAMESPACE_BEGIN
