+++ src/liblx/xml/XMLNamespaces.h
@@ -64,8 +64,8 @@
 #ifndef XMLNamespaces_h
 #define XMLNamespaces_h
 
-#include <liblx/xml/common/extern.h>
-#include <liblx/xml/common/liblxfwd.h>
+#include <liblx/common/extern.h>
+#include <liblx/common/liblxfwd.h>
 
 
 #ifdef __cplusplus
