+++ src/liblx/xml/XMLNode.h
@@ -151,9 +151,9 @@ else:
 #ifndef XMLNode_h
 #define XMLNode_h
 
-#include <liblx/xml/common/extern.h>
+#include <liblx/common/extern.h>
 #include <liblx/xml/XMLToken.h>
-#include <liblx/xml/common/liblxfwd.h>
+#include <liblx/common/liblxfwd.h>
 
 
 #ifdef __cplusplus
