+++ src/liblx/xml/XMLToken.h
@@ -117,14 +117,14 @@
 #ifndef XMLToken_h
 #define XMLToken_h
 
-#include <liblx/xml/common/extern.h>
+#include <liblx/common/extern.h>
 #include <liblx/xml/XMLAttributes.h>
 /** @cond doxygenLibsbmlInternal */
 #include <liblx/xml/XMLNamespaces.h>
 #include <liblx/xml/XMLOutputStream.h>
 /** @endcond */
 #include <liblx/xml/XMLToken.h>
-#include <liblx/xml/common/liblxfwd.h>
+#include <liblx/common/liblxfwd.h>
 
 
 #ifdef __cplusplus
