+++ src/liblx/xml/XMLTokenizer.h
@@ -49,7 +49,7 @@
 
 #include <deque>
 
-#include <liblx/xml/common/extern.h>
+#include <liblx/common/extern.h>
 #include <liblx/xml/XMLHandler.h>
 
 LIBLX_CPP_NAMESPACE_BEGIN
