+++ src/liblx/xml/XMLTriple.h
@@ -75,8 +75,8 @@
 #ifndef XMLTriple_h
 #define XMLTriple_h
 
-#include <liblx/xml/common/extern.h>
-#include <liblx/xml/common/liblxfwd.h>
+#include <liblx/common/extern.h>
+#include <liblx/common/liblxfwd.h>
 
 
 #ifdef __cplusplus
