+++ src/liblx/xml/sbmlStubs.h
@@ -11,9 +11,9 @@
 #include <ctype.h>
 #include <cstddef>
 
-#include <liblx/xml/common/extern.h>
+#include <liblx/common/extern.h>
 
-#include <liblx/xml/common/extern.h>
+#include <liblx/common/extern.h>
 
 
 LIBLX_CPP_NAMESPACE_BEGIN
